Revert "SonarQube clean up of code smells (#260)"
This reverts commit 767c909f81b6f41cafc685f1fbb1131f580a4169.

+++ src/EPR.Calculator.Service.Function/Converter/DecimalPrecisionConverter.cs
@@ -29,7 +29,7 @@ namespace EPR.Calculator.Service.Function.Converter
             }
             else
             {
-                writer.WriteRawValue(value.ToString(
+                writer.WriteRawValue(((decimal)value).ToString(
                     $"N{_precision}",
                     new NumberFormatInfo
                     {
